// src/app/interfaces/interface.ts
export interface User {
    firstName: string;
    lastName: string;
    companyName: string;
    designation: string;
    email: string;
    mobile: string;
    _id: string;
    submittedAt: Date;
}

export interface Industry {
    value: string;
    viewValue: string;
}

export interface Country {
    value: string;
    viewValue: string;
}

export interface State {
    value: string;
    viewValue: string;
}

export interface EmployeeSize {
    value: string;
    viewValue: string;
}

export interface Month {
    value: string;
    viewValue: string;
}

export interface Week {
    value: string;
    viewValue: string;
}

export interface ContactAgain {
    value: string;
    viewValue: string;
}

export interface Budget {
    value: string;
    viewValue: string;
}

export interface TechSolution {
    value: string;
    viewValue: string;
}

export interface BuyingProcess {
    value: string;
    viewValue: string;
}